fix: 드론 부품 조회 API 통합
교체 부품 리스트 조회 API & 재활용 부품 리스트 조회 API 로직 통합 구현

+++ src/main/java/com/weflo/backend/domain/component/repository/DroneComponentRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DroneComponentRepository extends JpaRepository<DroneComponent, Long> {
-    List<DroneComponent> findByDroneIdAndPointGreaterThanEqual(Long droneId, int point);
+    List<DroneComponent> findByDroneIdAndPointGreaterThanEqual(Long droneId, Long point);
 }
